Split fixers into several classes
