+++ src/Generation/Generator/Fixer/Bitfields.cs
@@ -1,34 +1,19 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
-using Generator.Model;
+using Generator.Fixer.Bitfield;
 
 namespace Generator.Fixer;
 
 public static class Bitfields
 {
+    private static readonly List<Fixer<GirModel.Bitfield>> Fixers = new()
+    {
+        new DisableDuplicateMembersFixer()
+    };
+
     public static void Fixup(IEnumerable<GirModel.Bitfield> bitfields)
     {
         foreach (var bitfield in bitfields)
-            Fixup(bitfield);
-    }
-
-    private static void Fixup(GirModel.Bitfield bitfield)
-    {
-        DisableDuplicateMembers(bitfield);
-    }
-
-    private static void DisableDuplicateMembers(GirModel.Bitfield bitfield)
-    {
-        foreach (var grouping in bitfield.Members.GroupBy(member => member.Name))
-        {
-            if (grouping.Count() <= 1)
-                continue;
-
-            foreach (var member in grouping.Skip(1)) //Disable all but the first member
-            {
-                Member.Disable(member);
-                Log.Debug($"{bitfield.Name}: Disabled member {member.Name} with value {member.Value} because there is another member with the same name");
-            }
-        }
+            foreach (var fixer in Fixers)
+                fixer.Fixup(bitfield);
     }
 }
