+++ src/Generation/Generator/Fixer/Fixer.cs
@@ -1,0 +1,6 @@
+﻿namespace Generator.Fixer;
+
+public interface Fixer<in T>
+{
+    void Fixup(T obj);
+}
